+++ esp-hal/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `rng` module has been rewritten (#3829)
 - Update `embassy-usb` to v0.5.0 (#3848)
 - `aes::Key` variants have been renamed from bytes to bits (e.g. `Key16 -> Key128`) (#3845)
+- `aes::Mode` has been replaced by `Operation`. The key length is now solely determined by the key. (#3882)
+- `Aes::process` has been split into `Aes::encrypt` and `Aes::decrypt` (#3882)
 
 ### Fixed
 
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `Trng::new` (replaced by `Trng::try_new`) (#3829)
+- `AesDma::write_block` has been removed. (#3882)
 
 ## [v1.0.0-rc.0] - 2025-07-16